dom/tests/mochitest/bugs/test_bug440572.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=440572
-->
<head>
  <title>Test for Bug 440572</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runtests()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=440572">Mozilla Bug 440572</a>

<script class="testbody" type="text/javascript">

/** Test for Bug 440572 **/

var messages = [];

function receiveMessage(e)
{
  is(e.origin, "http://example.org", "wrong sender!");
  messages.push(e.data);
}

window.addEventListener("message", receiveMessage, false);

function runtests()
{
  is(messages.length, 3, "received the right number of messages.");
  is(messages[0], "success", "test in frame failed.");
  isnot(messages[1], "success", "parent[\"content\"] should be the WebIDL property of Window.");
  isnot(messages[2], "success", "parent[\"dump\"] should be the WebIDL property of Window.");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
<br>
<iframe name="test" src="http://example.org:80/tests/dom/tests/mochitest/bugs/iframe_bug440572.html"></iframe>
<br>
<iframe name="content" src="http://example.org:80/tests/dom/tests/mochitest/bugs/iframe_bug440572.html"></iframe>
<br>
<iframe name="dump" src="http://example.org:80/tests/dom/tests/mochitest/bugs/iframe_bug440572.html"></iframe>
</body>
</html>
